backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.1
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.5
django-appconf==1.0.6
django-celery-beat==2.7.0
django-celery-email==3.0.0
django-cors-headers==4.6.0
django-filter==24.3
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
flower==2.0.1
humanize==4.11.0
idna==3.10
kombu==5.4.2
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
qrcode==8.0
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
